requirements.txt
discord.py
find.py